build: Move TCTI code from tcti/ to src/tss2-tcti/
Minor fixups were required to 'include' statements in a few test cases
as well as the pkgconfig files.
