--- lib/tcti-mssim.pc.in
@@ -1,7 +1,0 @@
-Name: tcti-socket
-Description: TCTI library for communicating with a TPM over a socket.
-URL: https://github.com/tpm2-software/tpm2-tss
-Version: @VERSION@
-Requires: marshal
-Cflags: -I@includedir@
-Libs: -ltcti-socket -L@libdir@
