+++ lib/tss2-tcti-device.pc.in
@@ -1,7 +1,7 @@
-Name: tcti-device
+Name: tss2-tcti-device
 Description: TCTI library for communicating with a TPM device node.
 URL: https://github.com/tpm2-software/tpm2-tss
 Version: @VERSION@
-Requires: marshal
+Requires: tss2-mu
 Cflags: -I@includedir@
-Libs: -ltcti-device -L@libdir@
+Libs: -ltss2-tcti-device -L@libdir@
