+++ lib/tss2-tcti-mssim.pc.in
@@ -1,0 +1,7 @@
+Name: tss2-tcti-mssim
+Description: TCTI library for communicating with the Microsoft TPM2 simulator.
+URL: https://github.com/tpm2-software/tpm2-tss
+Version: @VERSION@
+Requires: tss2-mu
+Cflags: -I@includedir@
+Libs: -ltss2-tcti-mssim -L@libdir@
