+++ src/tss2-tcti/tcti-device.c
