+++ test/integration/main-esapi.c
@@ -9,7 +9,7 @@
 #include "esys_types.h"
 #include "esys_iutil.h"
 #include "tss2_tcti.h"
-#include "tcti/tcti.h"
+#include "tss2-tcti/tcti.h"
 
 /** Define a proxy tcti that returns yielded on every second invocation
  * thus the corresponding handling code in ESAPI can be tested.
